temp analysis tool
executes "PDSTEP_analyze.exe" reads all of the text exports, and
performs analysis on the data. Half-complete: joint angle values are all
of the same - weird, needs more attention.
* Adds left foot COM trace to overall COM trace plots.
* Contact force data is unreliable (very noisy)
* Touch sensor might be used to establish the length of the step
(however, the foot goes in and out of contact when it is visibly on the
ground - a quirk of BULLET, haven't found a way to have constant
collision detection method)